moved benchmarking to separate class
